Fix warnings in rpc-fast, needle container, super container, and cfs namespace
- rpc.h: replace empty-body while(cond); spinlocks with while(cond) {} to silence -Wempty-body
- rpc.cc: cast myturn to unsigned to fix -Wsign-compare vs unsigned tid
- needle/common.h: restore session param name in Make() (was wrongly commented out while still used); silence unused session/off in Write/Read
- super/common.h: silence unused session params in root/set_root/freelist/set_freelist (params required by interface but not used in body)
- cfs/server/namespace.cc: initialize ip=NULL to suppress -Wmaybe-uninitialized (else branch leaves ip unset)

+++ libfs/src/bcs/backend/rpc-fast/rpc.cc
@@ -834,7 +834,7 @@ void* rpcs::rpc_server_kernel(void* arg) {
  	curr != rpc_reg.rpc_queue.end();
 	curr++, myturn++) {
    
-      if( myturn % NUMTHREADS != tid )
+      if( (unsigned)myturn % NUMTHREADS != tid )
 	continue;
       
       s_ce = *curr;
